test: refactor pool tests
